+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 2026.8.16 - 2026-08-16
 
 ### Fixed